requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
psycopg2